first pause in video 2
