--- gymbuddies/KavSoftContentView.swift
@@ -1,20 +1,0 @@
-//
-//  ListMessagesView.swift
-//  gymbuddies
-//
-//  Created by Kareha on 2/2/21.
-//
-
-import SwiftUI
-
-struct KavSoftContentView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct KavSoftContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        KavSoftContentView()
-    }
-}
