+++ gymbuddies/MessagesView.swift
@@ -69,61 +69,8 @@ struct MessagesView_Previews: PreviewProvider {
     }
 }
 
-class MainObservable : ObservableObject{
-    
-    @Published var recents = [Recent]()
-    @Published var norecetns = false
-    @ObservedObject private var userData = getUserData()
-    
-    init() {
-        
-        let db = Firestore.firestore()
-        let uid = Auth.auth().currentUser?.uid
-        
-        db.collection("users").document(uid!).collection("recents").order(by: "date", descending: true).addSnapshotListener { (snap, err) in
-            
-            if err != nil{
-                
-                print((err?.localizedDescription)!)
-                self.norecetns = true
-                return
-            }
-            
-            for i in snap!.documentChanges{
-                
-                let id = i.document.documentID
-                let name = i.document.get("name") as! String
-                let pic = i.document.get("pic") as! String
-                let lastmsg = i.document.get("lastmsg") as! String
-                let stamp = i.document.get("date") as! Timestamp
-                
-                let formatter = DateFormatter()
-                formatter.dateFormat = "dd/MM/yy"
-                let date = formatter.string(from: stamp.dateValue())
-                
-                formatter.dateFormat = "hh:mm a"
-                let time = formatter.string(from: stamp.dateValue())
-                
-                self.recents.append(Recent(id: id, name: name, pic: pic, lastmsg: lastmsg, time: time, date: date, stamp: stamp.dateValue()))
-                
-            }
-            
-            
-        }
-    }
-    
-}
 
-struct Recent : Identifiable {
-    
-    var id : String
-    var name : String
-    var pic : String
-    var lastmsg : String
-    var time : String
-    var date : String
-    var stamp : Date
-}
+
 
 struct RecentCellView : View {
     var url : String
